requirements.txt
fastapi[standard]
sqlmodel    
psycopg
psycopg-binary
pydantic-settings